requirements.txt
# python 3.6
cudatoolkit
elasticdeform
fvcore
iopath
matplotlib
numpy
opencv-python
pandas
pydicom
pytorch
pyyaml
scikit-image==0.18.0rc1
scikit-learn
scipy
SimpleITK
tensorboard
torchaudio
torchvision
tqdm
sacred
pymongo
